Fix issues with running `lein ring server`
1) Compatibility issue: lein-ring and latest leiningen

```
clojure.lang.Compiler$CompilerException: Syntax error macroexpanding clojure.core/fn at (clojure/core/unify.clj:83:18).
```
`lein-ring` seems to be uncompatible with Leiningen 2.9.0 and above.

Similar issues:
* https://github.com/Azure/draft/issues/932
* https://translate.google.com/translate?hl=en&sl=zh-CN&u=https://xbuba.com/questions/54810552&prev=search

2) Incorrect lein-ring handler
lein-ring had a wrong reference, but even if it was correct it should to be a handler not a fn.
Since evaluation of app at boot doesn't seem to be harmful, I let myself to change it.